Fix gulpfile.js
Fixes order of javascript files and reverts to myth from sass.
