+++ package.json
@@ -32,14 +32,13 @@
     "gulp-cssnano": "^2.1.1",
     "gulp-jshint": "^2.0.0",
     "gulp-livereload": "^3.8.1",
+    "gulp-myth": "^1.1.0",
     "gulp-notify": "^2.2.0",
     "gulp-plumber": "^1.1.0",
     "gulp-rename": "^1.2.2",
-    "gulp-sass": "^2.2.0",
     "gulp-sourcemaps": "^1.6.0",
     "gulp-uglify": "^1.5.2",
     "gulp-watch": "^4.3.5",
-    "jshint": "^2.9.1",
-    "sassdoc": "^2.1.20"
+    "jshint": "^2.9.1"
   }
 }
